+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/includes/autowiring_classes_80.php
@@ -4,7 +4,7 @@ namespace Symfony\Component\DependencyInjection\Tests\Compiler;
 
 use Symfony\Component\DependencyInjection\Attribute\AsDecorator;
 use Symfony\Component\DependencyInjection\Attribute\Autowire;
-use Symfony\Component\DependencyInjection\Attribute\MapDecorated;
+use Symfony\Component\DependencyInjection\Attribute\AutowireDecorated;
 use Symfony\Component\DependencyInjection\Attribute\TaggedIterator;
 use Symfony\Component\DependencyInjection\Attribute\TaggedLocator;
 use Symfony\Component\DependencyInjection\ContainerInterface;
@@ -80,7 +80,7 @@ class AsDecoratorFoo implements AsDecoratorInterface
 #[AsDecorator(decorates: AsDecoratorFoo::class, priority: 10)]
 class AsDecoratorBar10 implements AsDecoratorInterface
 {
-    public function __construct(string $arg1, #[MapDecorated] AsDecoratorInterface $inner)
+    public function __construct(string $arg1, #[AutowireDecorated] AsDecoratorInterface $inner)
     {
     }
 }
@@ -88,7 +88,7 @@ class AsDecoratorBar10 implements AsDecoratorInterface
 #[AsDecorator(decorates: AsDecoratorFoo::class, priority: 20)]
 class AsDecoratorBar20 implements AsDecoratorInterface
 {
-    public function __construct(string $arg1, #[MapDecorated] AsDecoratorInterface $inner)
+    public function __construct(string $arg1, #[AutowireDecorated] AsDecoratorInterface $inner)
     {
     }
 }
@@ -96,7 +96,7 @@ class AsDecoratorBar20 implements AsDecoratorInterface
 #[AsDecorator(decorates: \NonExistent::class, onInvalid: ContainerInterface::NULL_ON_INVALID_REFERENCE)]
 class AsDecoratorBaz implements AsDecoratorInterface
 {
-    public function __construct(#[MapDecorated] AsDecoratorInterface $inner = null)
+    public function __construct(#[AutowireDecorated] AsDecoratorInterface $inner = null)
     {
     }
 }
@@ -106,7 +106,7 @@ class AutowireNestedAttributes implements AsDecoratorInterface
 {
     public function __construct(
         #[Autowire([
-            'decorated' => new MapDecorated(),
+            'decorated' => new AutowireDecorated(),
             'iterator' => new TaggedIterator('foo'),
             'locator' => new TaggedLocator('foo'),
             'service' => new Autowire(service: 'bar')